* master: (589 commits)
  Adding delete user flow. About time, since I'm sick of doing this myself.
  Deractivating premiums.
  Adding a lock for the intelligence slider for focus stories. If the user selects focus and there are actually focus stories, lock it for the future when there are no unread focus stories.
  Uppiung the find next unread story titles page limit from 12 to 50.
  Hiding folders with no unread stories in Unread/Focus mode.
  Adding autocomplete to emailing a story dialog.
  Using queue for scheduling feed fetches on load.
  Renaming celerybeat supervisor task.
  Fixing insta-fetching.
  Attempting to correctly save duplicate feeds.
  Adding tons of logging (especially user-facing) to feed exception handling. Much much better feed address and link fetch history.
  New enterprise ios build includes full training.
  Changing intelligence training on social feeds to apply to original feed as well.
  Using a popover for training on the iPad. Also adding the train button to the bottom of every story.
  Finishing training on iPad. Works in river too. Just need to test social and it's shipping.
  Fixing title selectin bug when the title was too short to register the selection loupe.
  Training feeds and stories in feeds. Needs testing on ipad, in river, and on social.
  Finish look and feel of ios training for both story and feed. Just need to send requests, redraw intelligence, recount and reclassify stories, and update scores. All already written to support inline training.
  A few more log color tweaks.
  Adding clean stories task. Updating log colors.
  ...

Conflicts:
	.gitignore
	apps/reader/views.py
	apps/social/views.py
	templates/social/social_page.xhtml
	templates/social/social_story.xhtml
	templates/social/story_comments.xhtml

+++ media/ios/ASI/ASIDataDecompressor.m
@@ -211,7 +211,7 @@
 
 + (NSError *)inflateErrorWithCode:(int)code
 {
-	return [NSError errorWithDomain:NetworkRequestErrorDomain code:ASICompressionError userInfo:[NSDictionary dictionaryWithObjectsAndKeys:[NSString stringWithFormat:@"Decompression of data failed with code %hi",code],NSLocalizedDescriptionKey,nil]];
+	return [NSError errorWithDomain:NetworkRequestErrorDomain code:ASICompressionError userInfo:[NSDictionary dictionaryWithObjectsAndKeys:[NSString stringWithFormat:@"Decompression of data failed with code %hi",(short)code],NSLocalizedDescriptionKey,nil]];
 }
 
 @synthesize streamReady;
